requirements.txt
python-telegram-bot==20.4
python-dotenv==1.0.0
requests==2.31.0
Flask==2.3.3
Werkzeug==2.3.7
gunicorn==21.2.0 